separated spec files for expand
