+++ cypress/integration/v2/blocks/expand/other-tests.spec.js
@@ -1,0 +1,19 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import {
+	blockExist,
+	blockError,
+	typeContent,
+} from '.'
+
+describe( 'Expand Block ( Other Tests )', registerTests( [
+	blockExist,
+	blockError,
+	typeContent,
+] ) )
